VAD refactoring
