+++ Sources/WhisperKit/Core/Audio/AudioChunker.swift
@@ -48,9 +48,9 @@ open class VADAudioChunker: AudioChunking {
     private let windowPadding: Int
     private let vad: VoiceActivityDetector
 
-    init(windowPadding: Int = 16000, vad: VoiceActivityDetector = EnergyVAD()) {
+    public init(windowPadding: Int = 16000, vad: VoiceActivityDetector? = nil) {
         self.windowPadding = windowPadding
-        self.vad = vad
+        self.vad = vad ?? EnergyVAD()
     }
 
     private func splitOnMiddleOfLongestSilence(audioArray: [Float], startIndex: Int, endIndex: Int) -> Int {
