+++ Sources/WhisperKit/Core/Audio/AudioStreamTranscriber.swift
