+++ Sources/WhisperKit/Core/Audio/EnergyVAD.swift
