+++ Sources/WhisperKit/Core/Text/SegmentSeeker.swift
