+++ Tests/WhisperKitTests/RegressionTests.swift
@@ -1,6 +1,6 @@
 import CoreML
 import Hub
-@testable import WhisperKit
+import WhisperKit
 import XCTest
 
 @available(macOS 13, iOS 16, watchOS 10, visionOS 1, *)
